+++ sway-core/src/type_system/type_info.rs
@@ -4,7 +4,7 @@ use crate::{
     Ident,
 };
 use sway_error::error::CompileError;
-use sway_types::{integer_bits::IntegerBits, span::Span, Spanned};
+use sway_types::{integer_bits::IntegerBits, span::Span};
 
 use std::{
     collections::HashSet,
@@ -26,28 +26,6 @@ impl fmt::Display for AbiName {
                 AbiName::Known(cp) => cp.to_string(),
             }),
         )
-    }
-}
-
-#[derive(Debug, Clone, Hash)]
-pub struct Length {
-    val: usize,
-    span: Span,
-}
-
-impl Length {
-    pub fn new(val: usize, span: Span) -> Self {
-        Length { val, span }
-    }
-
-    pub fn val(&self) -> usize {
-        self.val
-    }
-}
-
-impl Spanned for Length {
-    fn span(&self) -> Span {
-        self.span.clone()
     }
 }
 
@@ -291,7 +269,7 @@ impl PartialEqWithTypeEngine for TypeInfo {
                         .as_deref()
                         .eq(&r_type_args.as_deref(), type_engine)
             }
-            (Self::Str(l), Self::Str(r)) => l.val == r.val,
+            (Self::Str(l), Self::Str(r)) => l.val() == r.val(),
             (Self::UnsignedInteger(l), Self::UnsignedInteger(r)) => l == r,
             (
                 Self::Enum {
@@ -351,7 +329,7 @@ impl PartialEqWithTypeEngine for TypeInfo {
                 type_engine
                     .look_up_type_id(l0.type_id)
                     .eq(&type_engine.look_up_type_id(r0.type_id), type_engine)
-                    && l1.val == r1.val
+                    && l1.val() == r1.val()
             }
             (TypeInfo::Storage { fields: l_fields }, TypeInfo::Storage { fields: r_fields }) => {
                 l_fields.eq(r_fields, type_engine)
@@ -375,7 +353,7 @@ impl DisplayWithTypeEngine for TypeInfo {
         let s = match self {
             Unknown => "unknown".into(),
             UnknownGeneric { name, .. } => name.to_string(),
-            Str(x) => format!("str[{}]", x.val),
+            Str(x) => format!("str[{}]", x.val()),
             UnsignedInteger(x) => match x {
                 IntegerBits::Eight => "u8",
                 IntegerBits::Sixteen => "u16",
@@ -425,7 +403,9 @@ impl DisplayWithTypeEngine for TypeInfo {
                         .unwrap_or_else(|| "None".into())
                 )
             }
-            Array(elem_ty, count) => format!("[{}; {}]", type_engine.help_out(elem_ty), count.val),
+            Array(elem_ty, count) => {
+                format!("[{}; {}]", type_engine.help_out(elem_ty), count.val())
+            }
             Storage { .. } => "contract storage".into(),
             RawUntypedPtr => "raw untyped ptr".into(),
             RawUntypedSlice => "raw untyped slice".into(),
@@ -546,7 +526,7 @@ impl TypeInfo {
         match self {
             Unknown => "unknown".into(),
             UnknownGeneric { name, .. } => name.to_string(),
-            Str(x) => format!("str[{}]", x.val),
+            Str(x) => format!("str[{}]", x.val()),
             UnsignedInteger(x) => match x {
                 IntegerBits::Eight => "u8",
                 IntegerBits::Sixteen => "u16",
@@ -578,7 +558,7 @@ impl TypeInfo {
                 format!("contract caller {}", abi_name)
             }
             Array(elem_ty, length) => {
-                format!("[{}; {}]", elem_ty.json_abi_str(type_engine), length.val)
+                format!("[{}; {}]", elem_ty.json_abi_str(type_engine), length.val())
             }
             Storage { .. } => "contract storage".into(),
             RawUntypedPtr => "raw untyped ptr".into(),
@@ -594,7 +574,7 @@ impl TypeInfo {
     ) -> CompileResult<String> {
         use TypeInfo::*;
         let name = match self {
-            Str(len) => format!("str[{}]", len.val),
+            Str(len) => format!("str[{}]", len.val()),
             UnsignedInteger(bits) => {
                 use IntegerBits::*;
                 match bits {
@@ -748,7 +728,7 @@ impl TypeInfo {
                     Some(name) => name,
                     None => return name,
                 };
-                format!("a[{};{}]", name, length.val)
+                format!("a[{};{}]", name, length.val())
             }
             RawUntypedPtr => "rawptr".to_string(),
             RawUntypedSlice => "rawslice".to_string(),
@@ -777,7 +757,7 @@ impl TypeInfo {
             TypeInfo::Tuple(fields) => fields
                 .iter()
                 .any(|field_type| id_uninhabited(field_type.type_id)),
-            TypeInfo::Array(elem_ty, length) => length.val > 0 && id_uninhabited(elem_ty.type_id),
+            TypeInfo::Array(elem_ty, length) => length.val() > 0 && id_uninhabited(elem_ty.type_id),
             _ => false,
         }
     }
@@ -826,7 +806,7 @@ impl TypeInfo {
                 all_zero_sized
             }
             TypeInfo::Array(elem_ty, length) => {
-                length.val == 0
+                length.val() == 0
                     || type_engine
                         .look_up_type_id(elem_ty.type_id)
                         .is_zero_sized(type_engine)
@@ -846,7 +826,7 @@ impl TypeInfo {
                     .can_safely_ignore(type_engine)
             }),
             TypeInfo::Array(elem_ty, length) => {
-                length.val == 0
+                length.val() == 0
                     || type_engine
                         .look_up_type_id(elem_ty.type_id)
                         .can_safely_ignore(type_engine)
@@ -1449,7 +1429,7 @@ impl TypeInfo {
                 type_engine
                     .look_up_type_id(l0.type_id)
                     .is_subset_of(&type_engine.look_up_type_id(r0.type_id), type_engine)
-                    && l1.val == r1.val
+                    && l1.val() == r1.val()
             }
             (
                 Self::Custom {
